pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-md-pdf"
version = "0.1.0"
description = "Model Context Protocol server for Markdown to Word/PDF conversion with template support"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Basel", email = "basel@example.com" }
]
keywords = ["mcp", "markdown", "pdf", "word", "docx", "converter", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=0.1.0",
    "pillow>=10.0.0",
    "python-docx>=1.1.0",
    "pywin32>=306; platform_system=='Windows'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "psutil>=5.9.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mcp-md-pdf = "md_pdf_mcp.server:main"

[project.urls]
Homepage = "https://github.com/yourusername/mcp-md-pdf"
Repository = "https://github.com/yourusername/mcp-md-pdf"
Issues = "https://github.com/yourusername/mcp-md-pdf/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
